% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg.R
\name{get_api_key}
\alias{get_api_key}
\title{Get API key for auth.}
\usage{
get_api_key(api_key = Sys.getenv("SENDGRID_API_KEY"))
}
\arguments{
\item{api_key}{SendGrid API key.}
}
\value{
api_key
}
\description{
Get API key for auth.
}
\keyword{internal}
