% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg.R
\name{is_email_address}
\alias{is_email_address}
\title{Check if a string is in a valid email format}
\usage{
is_email_address(email)
}
\arguments{
\item{email}{A character string.}
}
\value{
A logical value.
}
\description{
Check if a string is in a valid email format
}
\examples{
sg:::is_email_address("valid.email@example.com")
sg:::is_email_address("another-valid-email@domain.org")
sg:::is_email_address("invalid-email@com")
sg:::is_email_address("invalid.email@domain")
}
\keyword{internal}
