% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.R
\name{predict.sg.SL}
\alias{predict.sg.SL}
\title{Predict method for sg.SL object}
\usage{
predict.sg.SL(object, newdata = NULL)
}
\arguments{
\item{object}{Fitted object from \code{sg.SL}.}

\item{newdata}{New W values for prediction.}
}
\value{
A matrix of conditional average treatment effects (estimated by SuperLearner) with rows corresponding to the different covariate values in the rows of \code{newdata} and columns corresponding to the different treatments.
}
\description{
Obtains predictions on a new data set from an sg.SL fit.
}
\details{
If \code{newdata} is omitted, then predicted values from \code{object} are returned.
}
\keyword{SuperLearner,}
\keyword{average}
\keyword{blip}
\keyword{conditional}
\keyword{effect}
\keyword{function}
\keyword{treatment}
