% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg.R
\name{is_email_address}
\alias{is_email_address}
\title{Check if a string is in a valid email format}
\usage{
is_email_address(email)
}
\arguments{
\item{email}{A character string.}
}
\value{
A logical value.
}
\description{
Check if a string is in a valid email format
}
\keyword{internal}
